Intrinsics
Experimental. Intrinsics act as builtins, but live in a file system folder, rather than baked
into veil (or node). LSE packaging does not built veil or node, so it cannot bake in its
builtins. LSE puts the builtins into a special folder, which is communicated to veil
through VEIL_INTRINSIC_PATH. LSE for node has a similar mechanism.

The packages are called intrinsics because they are not normal NPM packages and
they do not have the same "baked in" properties of builtin packages.